Fix printing of the .git path when verbosity is enabled
When enabling verbose output (`-v`), the first line printed is the `git`
directory being used to look up the git metadata. In `1.1.0` of autotag there is
an issue where the path has an extra space before it, as well as an extra `/`.

```
2017/05/11 00:22:51 Opening repo at  .//.git
```

This change adds a `generateGitDirPath()` function which takes the `repoPath`
and calculates the absolute path to its `.git` directory. This is then used in
the debug log line, as well as in opening the repository. The absolute path
being printed also seems to fit well with verbosity being turned-up.

In addition to the above change, an erroneous space was removed from the
printing of the log line. When using `fmt.Println()`, arguments separated by a
comma (`,`) automatically have a space added between them when the line is
printed.

Here is how it looks now:

```
2017/05/11 00:39:14 Opening repo at /Users/theckman/go/src/github.com/pantheon-systems/autotag/.git
```