Document how the tagging calculations are done in the README.md file
This change adds additional information to the `README.md` file to explain how
the utility calculates tags. This includes information about needing to create
an initial `v0.0.0` tag, as well as how to bump the Major and Minor versions
with information provided in your commit messages.

This should provide enough information for developers to understand how
`autotag` works and to confidently use it.

Fixes #9