Fix incorrect information in README about how versions are bumped
In commit ce8963145 I documented how the tagging mechanism in `autotag` works. I
incorrectly stated that the markers for telling `autotag` how to increase the
version could be in any section of the commit message. This was incorrect. The
markers must be in the summary line of the git commit message, otherwise they
won't be observed by `autotag`.

This commit fixes the incorrect documentation.

Signed-off-by: Tim Heckman <t@heckman.io>