update version that imports lowercased unknwon/com
Before this commit, the autotag module depended on
github.com/gogits/git-module (now renamed to github.com/gogs/git-module).
Updating to gogs/git-module v0.8.3 which is shortly after that module
adopted go-modules, fixes also the dependency on Unknown/com, which in
turn removes the need for a replace directive.

Replace directives are incompatible with the future go install in 1.16
as they are not honored.

As this replace was seemingly only used to fix the casing issue,
updating to a more recent version seems to be the right approach.

A newer version of git-module (v1.1.3) was tried, but autotag uses APIs
that have been modified significantly in that release and would incur
more work than this commit.