Avoid writing error to Stdout
In order to run this in a script, and account for a case where the tool doesn't handle (i.e. a repo without tags yet), I had hoped to pipe the output to Stderr and assign it a default value if it errors out, but that proved difficult, given that the error output was sent to Stdout. This should mitigate that.